+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.0] - 2024-09-21
+
+### New Features
+
+* Language server arguments to ease integration with Neovim
+* Subcommand to generate shell completions
+
+### Removed Features
+
+* Shell completions will not be packaged with each release since they can be generated on the fly
+
 ## [3.2.0] - 2024-09-15
 
 ### Fixes